refactor(adapter): extract shared types into adapter_types.py
Decouple shared RuntimeAdapter types from tmux_adapter.py so the
upcoming SubprocessAdapter can import them without pulling in tmux
dependencies. adapter_protocol.py re-exports for backward compat.

Dispatch-ID: 20260406-160001-f28-pr0-adapter-types-C
