+++ scripts/lib/headless_transport_adapter.py
@@ -21,7 +21,7 @@ import shlex
 import subprocess
 from typing import Any, Dict, List, Optional
 
-from tmux_adapter import (
+from adapter_types import (
     CAPABILITY_DELIVER,
     CAPABILITY_HEALTH,
     CAPABILITY_INSPECT,
